chore(clippy): clear all warnings and gate clippy in CI (#739)
* chore(clippy): allow three workspace lints

Clippy is about to become a merge gate. too_many_arguments and
type_complexity describe API shape, not defects, and `!(a < b)` is
deliberate NaN handling that must not be rewritten to `>=`. Each
crate inherits the table with `[lints] workspace = true`.

* chore(clippy): apply machine-applicable fixes

cargo clippy --fix rewrote the suggestions it marks as safe:
needless borrows, unit lets, map_or forms, and nested ifs as
let chains. Comparisons and NaN handling are untouched. One
collapsible_match suggestion does not compile and stays for
a targeted allow.

* chore(clippy): clear the remaining warnings

The machine-applicable pass left local rewrites and a few
suggestions that would change a signature, drop a wire-format
comment, or collapse a match that does not compile. Those last
ones are allowed on the item, with the reason beside the attribute.

* chore(ci): require clippy -D warnings

Pull requests were not linted, so warnings accumulated until
-D warnings failed on the first crate. The new Windows job runs
the same clippy command developers must pass, and the CI contract
test reddens if that command is weakened or the job is demoted.

* chore(clippy): correct two allow reasons

CoreTrack::averaged checked the divisor by hand and claimed
checked_div would report zero. checked_div returns None for a
zero sample count, which is the unknown state, and the same
quotient otherwise. The hotkeys allow stays; its comment now
says the suggested collapse does not compile.

+++ crates/moon-core/src/db/analytics/time_zone.rs
@@ -86,11 +86,11 @@ pub(crate) fn install(conn: &Connection, axis: &ReportAxis) -> rusqlite::Result<
     // by two conversions cancelling.
     conn.create_scalar_function("mt_core_minute_of_day", 1, flags, move |ctx| {
         let secs = ctx.get::<i64>(0)?;
-        Ok(crate::util::display_time::minute_of_day(secs, chrono_tz::UTC).unwrap_or(0) as i64)
+        Ok(crate::util::display_time::minute_of_day(secs, chrono_tz::UTC).unwrap_or(0))
     })?;
     conn.create_scalar_function("mt_core_minute_of_week", 1, flags, move |ctx| {
         let secs = ctx.get::<i64>(0)?;
-        Ok(crate::util::display_time::minute_of_week(secs, chrono_tz::UTC).unwrap_or(0) as i64)
+        Ok(crate::util::display_time::minute_of_week(secs, chrono_tz::UTC).unwrap_or(0))
     })?;
 
     Ok(())
